完成任务页面修改

+++ api/resource.py
@@ -1,7 +1,7 @@
 # -*- coding: UTF-8 -*-
 __author__ = 'Joynice'
 from flask_restful import Resource
-from apps.cms.models import Task
+from apps.cms.models import Task, User
 from models import Asset
 from . import field
 from .parse import HostServerPost_parse, HostServerDelete_parse, HostServerGet_parse, HostServerUpgrade_parse, \
@@ -18,14 +18,21 @@ class HostServer(Resource):
         '''
         args = HostServerGet_parse.parse_args()
         task_id = args.get('task_id')
-        if not task_id:
-            return field.params_error(message='参数缺失')
-        task = Task.query.filter_by(task_id=task_id).first()
-        if task:
-            return field.success(message='获取状态成功！',
-                                 data={'task_id': task_id, 'url': task.url, 'cycle': task.cycle, 'number': task.number,
-                                       'state': task.state, 'result_id': task.result_id})
-
+        secret_key = args.get('secret_key')
+        if secret_key:
+            user = User.query.filter_by(secret_key=secret_key).first()
+            if user:
+                if not task_id:
+                    return field.params_error(message='参数缺失')
+                task = Task.query.filter_by(task_id=task_id).first()
+                if task:
+                    return field.success(message='获取状态成功！',
+                                         data={'task_id': task_id, 'url': task.url, 'cycle': task.cycle, 'number': task.number,
+                                               'state': task.state, 'result_id': task.result_id})
+            else:
+                return field.unauth_error(message='身份验证失败！')
+        else:
+            return field.params_error(message='身份验证参数缺失')
     def post(self):
         '''
         下发任务
@@ -37,21 +44,28 @@ class HostServer(Resource):
         url = args.get('url')
         cycle = args.get('cycle')
         number = args.get('number')
-        if url and cycle and number:
-            task = Task(url=url, cycle=str(cycle), number=number)
-            db.session.add(task)
-            try:
-                db.session.commit()
-            except Exception as e:
-                return field.params_error(message='创建任务失败!')
-            task_id = task.task_id
-            result_id = task.result_id
-            if number == 1:
-                WebCms(desurl=url).RunIt()
-                print(1111111111111111111111111111111111111)
-            return field.success(message='任务成功，结果请自己查询', data={'task_id': task_id, 'result_id': result_id})
+        secret_key = args.get('secret_key')
+        if secret_key:
+            user = User.query.filter_by(secret_key=secret_key).first()
+            if user:
+                if url and cycle and number:
+                    task = Task(url=url, cycle=str(cycle), number=number, user_id=user.id)
+                    db.session.add(task)
+                    try:
+                        db.session.commit()
+                    except Exception as e:
+                        return field.params_error(message='创建任务失败!')
+                    task_id = task.task_id
+                    result_id = task.result_id
+                    if number == 1:
+                        WebCms(desurl=url).RunIt()
+                    return field.success(message='任务成功，结果请自己查询', data={'task_id': task_id, 'result_id': result_id})
+                else:
+                    return field.params_error(message='参数缺失')
+            else:
+                return field.unauth_error(message='身份验证失败！')
         else:
-            return field.params_error(message='参数缺失')
+            return field.params_error(message='身份验证参数缺失！')
 
     def put(self):
         '''
@@ -63,20 +77,28 @@ class HostServer(Resource):
         url = args.get('url')
         cycle = args.get('cycle')
         number = args.get('number')
-        if not task_id:
-            return field.params_error(message='请传入更新ID')
-        if url or cycle or number:
-            task = Task.query.filter_by(task_id=task_id).first()
-            if task:
-                task.url = url or task.url
-                task.number = number or task.number
-                task.cycle = cycle or task.cycle
-                db.session.commit()
-                return field.success(message='更新任务成功！')
+        secret_key = args.get('secret_key')
+        if secret_key:
+            user = User.query.filter_by(secret_key=secret_key).first()
+            if user:
+                if not task_id:
+                    return field.params_error(message='请传入更新ID')
+                if url or cycle or number:
+                    task = Task.query.filter_by(task_id=task_id).first()
+                    if task:
+                        task.url = url or task.url
+                        task.number = number or task.number
+                        task.cycle = cycle or task.cycle
+                        db.session.commit()
+                        return field.success(message='更新任务成功！')
+                    else:
+                        return field.params_error(message='没有找到该任务')
+                else:
+                    return field.params_error(message='参数缺失')
             else:
-                return field.params_error(message='没有找到该任务')
+                return field.unauth_error(message='身份验证失败！')
         else:
-            return field.params_error(message='参数缺失')
+            return field.params_error(message='身份验证参数缺失！')
 
     def delete(self):
         '''
@@ -85,32 +107,48 @@ class HostServer(Resource):
         '''
         args = HostServerDelete_parse.parse_args()
         task_id = args.get('task_id')
-        if not task_id:
-            return field.params_error(message='参数缺少')
-        task = Task.query.filter_by(task_id=task_id).first()
-        if task:
-            db.session.delete(task)
-            db.session.commit()
-            return field.success(message='删除任务成功')
+        secret_key = args.get('secret_key')
+        if secret_key:
+            user = User.query.filter_by(secret_key=secret_key).first()
+            if user:
+                if not task_id:
+                    return field.params_error(message='参数缺少')
+                task = Task.query.filter_by(task_id=task_id).first()
+                if task:
+                    db.session.delete(task)
+                    db.session.commit()
+                    return field.success(message='删除任务成功')
+                else:
+                    return field.params_error(message='没有查到任务')
+            else:
+                return field.unauth_error(message='身份验证失败！')
         else:
-            return field.params_error(message='没有查到任务')
+            return field.params_error(message='身份验证参数缺失！')
+
 
 
 class Result(Resource):
     '''
     获取结果
     '''
-
     def get(self):
         args = ResultGet_parse.parse_args()
         result_id = args.get('result_id')
-        if not result_id:
-            return field.params_error('参数缺失')
-        result1 = Task.query.filter_by(result_id=result_id).first()
-        if result1:
-            result_data = result1.result
-            return field.success(message='查询成功',
-                                 data={'task_id': result1.task_id, 'result_id': result_id, 'result': result_data})
+        secret_key = args.get('secret_key')
+        if secret_key:
+            user = User.query.filter_by(secret_key=secret_key).first()
+            if user:
+                if not result_id:
+                    return field.params_error('参数缺失')
+                result1 = Task.query.filter_by(result_id=result_id).first()
+                if result1:
+                    result_data = result1.result
+                    return field.success(message='查询成功',
+                                         data={'task_id': result1.task_id, 'result_id': result_id, 'result': result_data})
+            else:
+                return field.unauth_error(message='身份验证失败！')
+        else:
+            return field.params_error(message='身份验证参数缺失！')
 
     def post(self):
         return field.method_error(message='不接受方法请求')
